Make `is_job_marked_acceptable_for_incident` usable from other commands
* Move `is_job_marked_acceptable_for_incident` to the openQA client so it
  can be used by other commands than the approver
* See https://progress.opensuse.org/issues/161267
